+++ R/traj_trajLevelMap.R
@@ -298,6 +298,8 @@ trajLevelMapStatic <-
            map.lty = 1,
            facet = NULL,
            ...) {
+    rlang::check_installed("ggplot2")
+
     # handle deprecated argument
     if (!is.null(facet)) {
       lifecycle::deprecate_soft(